requirements.txt
elevenlabs[pyaudio]==1.54.0
python-dotenv==1.0.1